remove intro fxml
